+++ src/TileDraw.java
@@ -7,10 +7,10 @@ public class TileDraw extends YldComponent implements YldGraphical {
 
     @Override
     public void draw(Graphics g) {
-        if(TileSystem.tiles != null) {
-            for(Tile tile : TileSystem.tiles) {
-                if(tile != null)
-                g.drawImage(tile.getTileImage(), tile.getX() + Camera.x, tile.getY() + Camera.y, Tile.getWidth(), Tile.getHeight(), null);
+        if (TileSystem.tiles != null) {
+            for (Tile tile : TileSystem.tiles) {
+                if (tile != null)
+                    g.drawImage(tile.getTileImage(), tile.getX() + (int) Camera.x, tile.getY() + (int) Camera.y, Tile.getWidth(), Tile.getHeight(), null);
             }
         }
     }
